rename multihash_factor to composition_factor
