+++ deepray/layers/embedding.py
@@ -15,13 +15,11 @@
 """Embedding layer."""
 import math
 import os
-from collections import defaultdict
-from typing import Dict, List
+from typing import Dict
 
 import numpy as np
 import pandas as pd
 import tensorflow as tf
-import tensorflow_recommenders_addons as tfra
 from absl import flags
 from keras import backend
 from keras import constraints
@@ -30,23 +28,13 @@ from keras import regularizers
 from keras.dtensor import utils
 from keras.engine import base_layer_utils
 from keras.engine.base_layer import Layer
-from keras.utils import tf_utils
-from tensorflow.python.distribute import distribution_strategy_context as distribute_ctx
-from tensorflow.python.distribute import values_util
-from tensorflow.python.eager import context
-from tensorflow.python.framework import ops
-from tensorflow.python.keras.utils import tf_utils
-from tensorflow.python.ops.variables import VariableAggregation
-from tensorflow.python.training.tracking import data_structures
-from tensorflow_recommenders_addons import dynamic_embedding as de
-from tensorflow_recommenders_addons.dynamic_embedding.python.ops import dynamic_embedding_variable as devar
-from tensorflow_recommenders_addons.dynamic_embedding.python.ops.dynamic_embedding_ops import DistributedVariableWrapper, TrainableWrapperDistributedPolicy
-from tensorflow.keras.layers import StringLookup
-from tensorflow.keras import layers
 from tensorflow import keras
+from tensorflow.keras import layers
+from tensorflow.keras.layers import StringLookup
+from tensorflow.python.keras.utils import tf_utils
 
 import deepray as dp
-from deepray.layers.bucketize import NumericaBucketIdLayer, Hash
+from deepray.layers.bucketize import Hash
 
 FLAGS = flags.FLAGS
 
@@ -540,385 +528,6 @@ class DualEmbeddingGroup(tf.keras.layers.Layer):
     self.cpu_embeddings.restore_checkpoint(checkpoint_path)
 
 
-class DynamicEmbedding(tf.keras.layers.Layer):
-  """
-  A keras style Embedding layer. The `Embedding` layer acts same like
-  [tf.keras.layers.Embedding](https://www.tensorflow.org/api_docs/python/tf/keras/layers/Embedding),
-  except that the `Embedding` has dynamic embedding space so it does
-  not need to set a static vocabulary size, and there will be no hash conflicts
-  between features.
-
-  The embedding layer allow arbirary input shape of feature ids, and get
-  (shape(ids) + embedding_size) lookup result. Normally the first dimension
-  is batch_size.
-
-  ### Example
-  ```python
-  embedding = dynamic_embedding.keras.layers.Embedding(8) # embedding size 8
-  ids = tf.constant([[15,2], [4,92], [22,4]], dtype=tf.int64) # (3, 2)
-  out = embedding(ids) # lookup result, (3, 2, 8)
-  ```
-
-  You could inherit the `Embedding` class to implement a custom embedding
-  layer with other fixed shape output.
-
-  TODO(Lifann) Currently the Embedding only implemented in eager mode
-  API, need to support graph mode also.
-  """
-
-  def __init__(
-      self,
-      embedding_size,
-      key_dtype=tf.int64,
-      value_dtype=tf.float32,
-      combiner='sum',
-      initializer=None,
-      devices=None,
-      name='DynamicEmbeddingLayer',
-      with_unique=True,
-      **kwargs
-  ):
-    """
-    Creates an Embedding layer.
-
-    Args:
-      embedding_size: An object convertible to int. Length of embedding vector
-        to every feature id.
-      key_dtype: Dtype of the embedding keys to weights. Default is int64.
-      value_dtype: Dtype of the embedding weight values. Default is float32
-      combiner: A string or a function to combine the lookup result. It's value
-        could be 'sum', 'mean', 'min', 'max', 'prod', 'std', etc. whose are
-        one of tf.math.reduce_xxx.
-      initializer: Initializer to the embedding values. Default is RandomNormal.
-      devices: List of devices to place the embedding layer parameter.
-      name: Name of the embedding layer.
-      with_unique: Bool. Whether if the layer does unique on `ids`. Default is True.
-
-      **kwargs:
-        trainable: Bool. Whether if the layer is trainable. Default is True.
-        bp_v2: Bool. If true, the embedding layer will be updated by incremental
-          amount. Otherwise, it will be updated by value directly. Default is
-          False.
-        restrict_policy: A RestrictPolicy class to restrict the size of
-          embedding layer parameter since the dynamic embedding supports
-          nearly infinite embedding space capacity.
-        init_capacity: Integer. Initial number of kv-pairs in an embedding
-          layer. The capacity will growth if the used space exceeded current
-          capacity.
-        partitioner: A function to route the keys to specific devices for
-          distributed embedding parameter.
-        kv_creator: A KVCreator object to create external KV storage as
-          embedding parameter.
-        max_norm: If not `None`, each values is clipped if its l2-norm is larger
-        distribute_strategy: Used when creating ShadowVariable.
-        keep_distribution: Bool. If true, save and restore python object with
-          devices information. Default is false.
-    """
-
-    try:
-      embedding_size = int(embedding_size)
-    except:
-      raise TypeError('embeddnig size must be convertible to integer, but get {}'.format(type(embedding_size)))
-
-    self.embedding_size = embedding_size
-    self.combiner = combiner
-    if initializer is None:
-      initializer = tf.keras.initializers.RandomNormal()
-    partitioner = kwargs.get('partitioner', devar.default_partition_fn)
-    trainable = kwargs.get('trainable', True)
-    self.max_norm = kwargs.get('max_norm', None)
-    self.keep_distribution = kwargs.get('keep_distribution', False)
-    self.with_unique = with_unique
-
-    parameter_name = name + '-parameter' if name else 'EmbeddingParameter'
-    with tf.name_scope('DynamicEmbedding'):
-      self.params = de.get_variable(
-          parameter_name,
-          key_dtype=key_dtype,
-          value_dtype=value_dtype,
-          dim=self.embedding_size,
-          devices=devices,
-          partitioner=partitioner,
-          shared_name='layer_embedding_variable',
-          initializer=initializer,
-          trainable=trainable,
-          checkpoint=kwargs.get('checkpoint', True),
-          init_size=kwargs.get('init_capacity', 0),
-          kv_creator=kwargs.get('kv_creator', None),
-          restrict_policy=kwargs.get('restrict_policy', None),
-          bp_v2=kwargs.get('bp_v2', False)
-      )
-
-      self.distribute_strategy = kwargs.get('distribute_strategy', None)
-      shadow_name = name + '-shadow' if name else 'ShadowVariable'
-      if distribute_ctx.has_strategy():
-        self.distribute_strategy = distribute_ctx.get_strategy()
-      if self.distribute_strategy:
-        strategy_devices = self.distribute_strategy.extended.worker_devices
-        self.shadow_impl = tf_utils.ListWrapper([])
-        for i, strategy_device in enumerate(strategy_devices):
-          with ops.device(strategy_device):
-            shadow_name_replica = shadow_name
-            if i > 0:
-              shadow_name_replica = "%s/replica_%d" % (shadow_name, i)
-            with context.device_policy(context.DEVICE_PLACEMENT_SILENT):
-              self.shadow_impl.as_list().append(
-                  de.shadow_ops.ShadowVariable(
-                      self.params,
-                      name=shadow_name_replica,
-                      max_norm=self.max_norm,
-                      trainable=trainable,
-                      distribute_strategy=self.distribute_strategy
-                  )
-              )
-      else:
-        self.shadow_impl = tf_utils.ListWrapper(
-            [de.shadow_ops.ShadowVariable(self.params, name=shadow_name, max_norm=self.max_norm, trainable=trainable)]
-        )
-    if len(self.shadow_impl.as_list()) > 1:
-      self._current_ids = data_structures.NoDependency([shadow_i.ids for shadow_i in self.shadow_impl.as_list()])
-      self._current_exists = data_structures.NoDependency([shadow_i.exists for shadow_i in self.shadow_impl.as_list()])
-      self.optimizer_vars = data_structures.NoDependency(
-          [shadow_i._optimizer_vars for shadow_i in self.shadow_impl.as_list()]
-      )
-    else:
-      self._current_ids = data_structures.NoDependency(self.shadow_impl.as_list()[0].ids)
-      self._current_exists = data_structures.NoDependency(self.shadow_impl.as_list()[0].exists)
-      self.optimizer_vars = self.shadow_impl.as_list()[0]._optimizer_vars
-    if distribute_ctx.has_strategy() and self.distribute_strategy and 'OneDeviceStrategy' not in str(
-        self.distribute_strategy
-    ) and not values_util.is_saving_non_distributed() and values_util.get_current_replica_id_as_int() is not None:
-      self.shadow = DistributedVariableWrapper(
-          self.distribute_strategy, self.shadow_impl.as_list(), VariableAggregation.NONE,
-          TrainableWrapperDistributedPolicy(VariableAggregation.NONE)
-      )
-    else:
-      self.shadow = self.shadow_impl.as_list()[0]
-    super(DynamicEmbedding, self).__init__(name=name, trainable=trainable, dtype=value_dtype)
-
-  def call(self, ids):
-    """
-    Compute embedding output for feature ids. The output shape will be (shape(ids),
-    embedding_size).
-
-    Args:
-      ids: feature ids of the input. It should be same dtype as the key_dtype
-        of the layer.
-
-    Returns:
-      A embedding output with shape (shape(ids), embedding_size).
-    """
-    ids = tf.convert_to_tensor(ids)
-    input_shape = tf.shape(ids)
-    embeddings_shape = tf.concat([input_shape, [self.embedding_size]], 0)
-    ids_flat = tf.reshape(ids, (-1,))
-    if self.with_unique:
-      with tf.name_scope(self.name + "/EmbeddingWithUnique"):
-        unique_ids, idx = tf.unique(ids_flat)
-        unique_embeddings = de.shadow_ops.embedding_lookup(self.shadow, unique_ids)
-        embeddings_flat = tf.gather(unique_embeddings, idx)
-    else:
-      embeddings_flat = de.shadow_ops.embedding_lookup(self.shadow, ids_flat)
-    embeddings = tf.reshape(embeddings_flat, embeddings_shape)
-    return embeddings
-
-  def get_config(self):
-    _initializer = self.params.initializer
-    if _initializer is None:
-      _initializer = tf.keras.initializers.Zeros()
-    _max_norm = None
-    if isinstance(self.max_norm, tf.keras.constraints.Constraint):
-      _max_norm = tf.keras.constraints.serialize(self.max_norm)
-
-    if self.params.restrict_policy:
-      _restrict_policy = self.params.restrict_policy.__class__
-    else:
-      _restrict_policy = None
-
-    config = {
-        'embedding_size': self.embedding_size,
-        'key_dtype': self.params.key_dtype,
-        'value_dtype': self.params.value_dtype,
-        'combiner': self.combiner,
-        'initializer': tf.keras.initializers.serialize(_initializer),
-        'devices': self.params.devices if self.keep_distribution else None,
-        'name': self.name,
-        'trainable': self.trainable,
-        'bp_v2': self.params.bp_v2,
-        'restrict_policy': _restrict_policy,
-        'init_capacity': self.params.init_size,
-        'partitioner': self.params.partition_fn,
-        'kv_creator': self.params.kv_creator if self.keep_distribution else None,
-        'max_norm': _max_norm,
-    }
-    return config
-
-
-class EmbeddingLayerGPU(DynamicEmbedding):
-
-  def call(self, ids):
-    with tf.name_scope(self.name + "/EmbeddingLookupUnique"):
-      ids_flat = tf.reshape(ids, [-1])
-      unique_ids, idx = tf.unique(ids_flat)
-      unique_embeddings = tfra.dynamic_embedding.shadow_ops.embedding_lookup(self.shadow, unique_ids)
-      embeddings_flat = tf.gather(unique_embeddings, idx)
-      embeddings_shape = tf.concat([tf.shape(ids), tf.constant(self.embedding_size, shape=(1,))], 0)
-      embeddings = tf.reshape(embeddings_flat, embeddings_shape)
-      return embeddings
-
-
-class EmbeddingLayerRedis(DynamicEmbedding):
-
-  def call(self, ids):
-    with tf.name_scope(self.name + "/EmbeddingLookupUnique"):
-      ids_flat = tf.reshape(ids, [-1])
-      with tf.device("/CPU:0"):
-        unique_ids, idx = tf.unique(ids_flat)
-      unique_embeddings = tfra.dynamic_embedding.shadow_ops.embedding_lookup(self.shadow, unique_ids)
-      embeddings_flat = tf.gather(unique_embeddings, idx)
-      embeddings_shape = tf.concat([tf.shape(ids), tf.constant(self.embedding_size, shape=(1,))], 0)
-      embeddings = tf.reshape(embeddings_flat, embeddings_shape)
-      return embeddings
-
-
-class CompositionalEmbedding(tf.keras.layers.Layer):
-  """
-  Compositional Embedding is designed for reducing Large-scale Sparse Embedding Weights.
-  See [Compositional Embeddings Using Complementary Partitions for Memory-Efficient Recommendation Systems]
-  (https://arxiv.org/abs/1909.02107)
-  """
-
-  def __init__(
-      self,
-      embedding_dim: int,
-      key_dtype: str,
-      value_dtype: str,
-      multihash_factor: str = "",
-      complementary_strategy: str = "Q-R",
-      operation: str = "add",
-      name: str = '',
-      devices=None,
-      **kwargs
-  ):
-    super().__init__(**kwargs)
-    self.devices = devices if devices else ['/GPU:0']
-    strategy_list = ["Q-R"]
-    op_list = ["add", "mul", "concat"]
-
-    if complementary_strategy not in strategy_list:
-      raise ValueError("The strategy %s is not supported" % complementary_strategy)
-    if operation not in op_list:
-      raise ValueError("The operation %s is not supported" % operation)
-    # if complementary_strategy == 'Q-R':
-    #   if num_of_partitions != 2:
-    #     raise ValueError("the num_of_partitions must be 2 when using Q-R strategy.")
-
-    self.embedding_dim = embedding_dim
-    self.key_dtype = key_dtype
-    self.value_dtype = value_dtype
-    self.multihash_factor = self.factor2decimal(multihash_factor) if multihash_factor else self.factor2decimal(
-        "16,16"
-    ) if self.key_dtype == "int32" else self.factor2decimal("32,32")
-    self.complementary_strategy = complementary_strategy
-    self.operation = operation
-    self.suffix = name
-
-  def factor2decimal(self, multihash_factor: str):
-    """
-    print(binary_to_decimal("16,16"))  # 4294901760, 65535
-    print(binary_to_decimal("17,15"))  # 4294934528, 32767
-    print(binary_to_decimal("15,17"))  # 4294836224, 131071
-    print(binary_to_decimal("14,18"))  # 4294705152, 262143
-    print(binary_to_decimal("13,19"))  # 4294443008, 524287
-
-    print(binary_to_decimal("32,32"))  # 18446744069414584320, 4294967295
-    print(binary_to_decimal("33,31"))  # 18446744071562067968, 2147483647
-    print(binary_to_decimal("31,33"))  # 18446744065119617024, 8589934591
-    print(binary_to_decimal("30,34"))  # 18446744056529682432, 17179869183
-    print(binary_to_decimal("29,35"))  # 18446744039349813248, 34359738367
-    """
-    if self.key_dtype == "int32":
-      Q, R = map(int, multihash_factor.split(','))
-      Q = (1 << Q) - 1 << (32 - Q)
-      R = (1 << R) - 1
-      return Q, R
-    elif self.key_dtype == "int64":
-      """
-  sign bit
-      1   111111111111111111111 111111111111111111111 111111111111111111111     -9223372036854776000
-      1   000000000000000000000 111111111111111111111 111111111111111111111     -9223367638808264705
-      1   111111111111111111111 000000000000000000000 111111111111111111111     -4398044413953
-      1   111111111111111111111 111111111111111111111 000000000000000000000     -2097151
-      """
-      Q, R, P = -9223367638808264705, -4398044413953, -2097152
-      return tf.constant(Q, dtype=tf.int64), tf.constant(R, dtype=tf.int64), tf.constant(P, dtype=tf.int64)
-    else:
-      raise ValueError(f"{self.key_dtype} type not support yet.")
-
-  def build(self, input_shape=None):
-    if not FLAGS.use_horovod:
-      self.multihash_emb = DynamicEmbedding(
-          embedding_size=self.embedding_dim,
-          key_dtype=self.key_dtype,
-          value_dtype=self.value_dtype,
-          initializer=tf.keras.initializers.GlorotUniform(),
-          name=f"embeddings_{self.suffix}/Compositional",
-          init_capacity=800000,
-          devices=self.devices,
-          kv_creator=de.CuckooHashTableCreator(saver=de.FileSystemSaver())
-      )
-    else:
-      import horovod.tensorflow as hvd
-      mpi_size = hvd.size()
-      mpi_rank = hvd.rank()
-      self.multihash_emb = de.keras.layers.HvdAllToAllEmbedding(
-          embedding_size=self.embedding_dim,
-          key_dtype=self.key_dtype,
-          value_dtype=self.value_dtype,
-          initializer=tf.keras.initializers.GlorotUniform(),
-          devices=self.devices,
-          init_capacity=800000,
-          name=f"embeddings_{self.suffix}/multihash",
-          kv_creator=de.CuckooHashTableCreator(saver=de.FileSystemSaver(proc_size=mpi_size, proc_rank=mpi_rank))
-      )
-
-  def call(self, inputs, *args, **kwargs):
-    if self.key_dtype == "int32":
-      ids_Q = tf.bitwise.bitwise_and(inputs, self.multihash_factor[0])
-      ids_R = tf.bitwise.bitwise_and(inputs, self.multihash_factor[1])
-      result_Q, result_R = tf.split(self.multihash_emb([ids_Q, ids_R]), num_or_size_splits=2, axis=0)
-      result_Q = tf.squeeze(result_Q, axis=0)
-      result_R = tf.squeeze(result_R, axis=0)
-      if self.operation == "add":
-        ret = tf.add(result_Q, result_R)
-        return ret
-      if self.operation == "mul":
-        ret = tf.multiply(result_Q, result_R)
-        return ret
-      if self.operation == "concat":
-        ret = tf.concat([result_Q, result_R], 1)
-        return ret
-    elif self.key_dtype == "int64":
-      ids_Q = tf.bitwise.bitwise_and(inputs, self.multihash_factor[0])
-      ids_R = tf.bitwise.bitwise_and(inputs, self.multihash_factor[1])
-      ids_P = tf.bitwise.bitwise_and(inputs, self.multihash_factor[2])
-      result_Q, result_R, result_P = tf.split(self.multihash_emb([ids_Q, ids_R, ids_P]), num_or_size_splits=3, axis=0)
-      result_Q = tf.squeeze(result_Q, axis=0)
-      result_R = tf.squeeze(result_R, axis=0)
-      result_P = tf.squeeze(result_P, axis=0)
-      if self.operation == "add":
-        ret = tf.add_n([result_Q, result_R, result_P])
-        return ret
-      if self.operation == "mul":
-        ret = tf.multiply(result_Q, result_R, result_P)
-        return ret
-      if self.operation == "concat":
-        ret = tf.concat([result_Q, result_R, result_P], 1)
-        return ret
-    else:
-      raise ValueError(f"{self.key_dtype} type not support yet.")
-
-
 class MaskedEmbeddingMean(tf.keras.layers.Layer):
 
   def __init__(self, hash_bucket_size, embedding_dim):
@@ -939,230 +548,6 @@ class MaskedEmbeddingMean(tf.keras.layers.Layer):
         tf.reduce_sum(embedding_mask_tensor, axis=1)
     )
     return tf.expand_dims(mean_tensor, axis=1)
-
-
-class DiamondEmbedding(tf.keras.layers.Layer):
-  """
-  Diamond Brother(金刚葫芦娃) has all the powers of the seven brothers, so should the Diamond Embedding too.
-  """
-
-  def __init__(self, feature_map: pd.DataFrame, fold_columns: Dict[str, List[str]], **kwargs):
-    super().__init__(**kwargs)
-    columns = ["bucket_boundaries", "hash_size", "voc_size", "multihash_factor", "storage_type"]
-    for col in columns:
-      if col not in feature_map.columns:
-        feature_map[col] = None
-
-    self.feature_map = feature_map
-    self.fold_columns = self.aggregate_by_dim(feature_map, fold_columns)
-
-  def aggregate_by_dim(self, df: pd.DataFrame, fold_columns: Dict[str, List[str]]) -> Dict[str, str]:
-    """
-    Aggregate the dim values for each group of names in the fold_columns list.
-
-    Args:
-        df (pd.DataFrame): The input DataFrame.
-        fold_columns (Dict[str, List[str]]): A list of lists of names to aggregate.
-
-    Returns:
-        Dict[str, str]: A dictionary containing the results of the aggregation.
-    """
-    folder_map = {}
-    for key, group in fold_columns.items():
-      dim_values = []
-      for name in group:
-        dim_value = df.loc[df['name'] == name]['dim'].values[0]
-        dim_values.append(dim_value)
-        folder_map[name] = key
-      if len(set(dim_values)) != 1:
-        raise ValueError(
-            f"Cannot aggregate {group} because dimensions are not equal. Names: {group}, Dims: {dim_values}"
-        )
-
-    # Record the remaining features that do not need to be folded
-    for name in self.feature_map[~(self.feature_map['ftype'].isin(["Label", "Weight"]))]["name"].values:
-      if name not in folder_map:
-        folder_map[name] = name
-    return folder_map
-
-  def build(self, input_shape):
-    self.embedding_layers = {}
-    self.hash_long_kernel = {}
-    self.numerical_bucket_kernel = {}
-    self.split_dims = defaultdict(list)
-    for name, length, dim, voc_size, dtype, hash_size, multihash_factor, storage_type, bucket_boundaries in self.feature_map[
-        ~(self.feature_map['ftype'].isin(["Label", "Weight"]))][[
-            "name", "length", "dim", "voc_size", "dtype", "hash_size", "multihash_factor", "storage_type",
-            "bucket_boundaries"
-        ]].values:
-
-      if self.is_valid_value(bucket_boundaries):
-        bucket_boundaries_list = sorted(set(map(float, bucket_boundaries.split(","))))
-        self.numerical_bucket_kernel[name] = NumericaBucketIdLayer(bucket_boundaries_list)
-
-      if self.is_valid_value(hash_size):
-        self.hash_long_kernel[name] = Hash(int(hash_size))
-        voc_size = int(hash_size)
-
-      if self.fold_columns[name] not in self.embedding_layers:
-        multihash_factor = self.feature_map.loc[self.feature_map['name'] == self.fold_columns[name]
-                                               ]['multihash_factor'].values[0] if self.fold_columns[
-                                                   name] in self.feature_map['name'].values else multihash_factor
-        storage_type = self.feature_map.loc[
-            self.feature_map['name'] == self.fold_columns[name]
-        ]['storage_type'].values[0] if self.fold_columns[name] in self.feature_map['name'].values else storage_type
-        if self.is_valid_value(multihash_factor):
-          self.embedding_layers[self.fold_columns[name]] = CompositionalEmbedding(
-              embedding_dim=dim,
-              key_dtype=tf.int32 if self.is_valid_value(bucket_boundaries) else dtype,
-              multihash_factor=multihash_factor,
-              operation="add",
-              name=self.fold_columns[name]
-          )
-        else:
-          if storage_type == "Redis":
-            redis_config = tfra.dynamic_embedding.RedisTableConfig(redis_config_abs_dir=FLAGS.config_file)
-            self.embedding_layers[self.fold_columns[name]] = EmbeddingLayerRedis(
-                embedding_size=dim,
-                key_dtype=tf.int32 if self.is_valid_value(bucket_boundaries) else dtype,
-                value_dtype=tf.float32,
-                initializer=tf.keras.initializers.GlorotUniform(),
-                name='embedding_redis' + self.fold_columns[name],
-                devices="/job:localhost/replica:0/task:0/CPU:0",
-                kv_creator=de.RedisTableCreator(redis_config)
-            )
-          elif storage_type == "DRAM":
-            devices = ['/CPU:0']
-            if not FLAGS.use_horovod:
-              self.embedding_layers[self.fold_columns[name]] = DynamicEmbedding(
-                  embedding_size=dim,
-                  key_dtype=tf.int32 if self.is_valid_value(bucket_boundaries) else dtype,
-                  value_dtype=tf.float32,
-                  initializer=tf.keras.initializers.GlorotUniform(),
-                  name='embedding_' + self.fold_columns[name],
-                  init_capacity=800000,
-                  devices=devices,
-                  kv_creator=de.CuckooHashTableCreator(saver=de.FileSystemSaver())
-              )
-            else:
-              import horovod.tensorflow as hvd
-
-              mpi_size = hvd.size()
-              mpi_rank = hvd.rank()
-              self.embedding_layers[self.fold_columns[name]] = de.keras.layers.HvdAllToAllEmbedding(
-                  # mpi_size=mpi_size,
-                  embedding_size=dim,
-                  key_dtype=tf.int32 if self.is_valid_value(bucket_boundaries) else dtype,
-                  value_dtype=tf.float32,
-                  initializer=tf.keras.initializers.GlorotUniform(),
-                  init_capacity=800000,
-                  name='embedding_' + self.fold_columns[name],
-                  devices=devices,
-                  kv_creator=de.CuckooHashTableCreator(
-                      saver=de.FileSystemSaver(proc_size=mpi_size, proc_rank=mpi_rank)
-                  )
-              )
-
-          else:
-            devices = ['/GPU:0']
-            if not FLAGS.use_horovod:
-              self.embedding_layers[self.fold_columns[name]] = DynamicEmbedding(
-                  embedding_size=dim,
-                  key_dtype=tf.int32 if self.is_valid_value(bucket_boundaries) else dtype,
-                  value_dtype=tf.float32,
-                  initializer=tf.keras.initializers.GlorotUniform(),
-                  name='embedding_' + self.fold_columns[name],
-                  init_capacity=800000,
-                  devices=devices,
-                  kv_creator=de.CuckooHashTableCreator(saver=de.FileSystemSaver())
-              )
-            else:
-              import horovod.tensorflow as hvd
-
-              mpi_size = hvd.size()
-              mpi_rank = hvd.rank()
-              self.embedding_layers[self.fold_columns[name]] = de.keras.layers.HvdAllToAllEmbedding(
-                  # mpi_size=mpi_size,
-                  embedding_size=dim,
-                  key_dtype=tf.int32 if self.is_valid_value(bucket_boundaries) else dtype,
-                  value_dtype=tf.float32,
-                  initializer=tf.keras.initializers.GlorotUniform(),
-                  init_capacity=800000,
-                  name='embedding_' + self.fold_columns[name],
-                  devices=devices,
-                  kv_creator=de.CuckooHashTableCreator(
-                      saver=de.FileSystemSaver(proc_size=mpi_size, proc_rank=mpi_rank)
-                  )
-              )
-
-      self.split_dims[self.fold_columns[name]].append(length)
-    # [1,1,1,10,1,1,30,1] -> [[3, 10, 2, 30, 1] and [False, True, False, True, False] for split sequence feature
-    self.split_dims_final = defaultdict(list)
-    self.is_sequence_feature = defaultdict(list)
-    tmp_sum = defaultdict(int)
-    for fold_name, dims in self.split_dims.items():
-      for dim in dims:
-        if dim == 1:
-          tmp_sum[fold_name] += 1
-        else:
-          if tmp_sum[fold_name] > 0:
-            self.split_dims_final[fold_name].append(tmp_sum[fold_name])
-            self.is_sequence_feature[fold_name].append(False)
-          self.split_dims_final[fold_name].append(dim)
-          self.is_sequence_feature[fold_name].append(True)
-          tmp_sum[fold_name] = 0
-
-    for fold_name, _sum in tmp_sum.items():
-      if _sum > 0:
-        self.split_dims_final[fold_name].append(_sum)
-        self.is_sequence_feature[fold_name].append(False)
-
-  def is_valid_value(self, x):
-    """
-    x1 = '-2.87,-1.93,-1.32,-0.84,-0.42,-0.01,0.4,0.86,1.52'
-    x2 = None
-    x3 = np.nan
-    x4 = ""
-    x5 = 6
-
-    print(is_valid_value(x1)) #  True
-    print(is_valid_value(x2)) #  False
-    print(is_valid_value(x3)) #  False
-    print(is_valid_value(x4)) #  False
-    print(is_valid_value(x5)) #  True
-    """
-    return isinstance(x, (int, str)) and bool(x)
-
-  def call(self, inputs, *args, **kwargs) -> Dict[str, List[tf.Tensor]]:
-    result = defaultdict(list)
-    id_tensors = defaultdict(list)
-    for code, name, hash_size, bucket_boundaries in self.feature_map[
-        ~(self.feature_map['ftype'].isin(["Label", "Weight"]))][["code", "name", "hash_size",
-                                                                 "bucket_boundaries"]].values:
-
-      input_tensor = inputs[name]
-      id_tensor_prefix_code = code << 47
-
-      if self.is_valid_value(bucket_boundaries):
-        input_tensor = self.numerical_bucket_kernel[name](input_tensor)
-
-      if self.is_valid_value(hash_size):
-        input_tensor = self.hash_long_kernel[name](input_tensor)
-
-      id_tensor = tf.bitwise.bitwise_or(input_tensor, id_tensor_prefix_code)
-      id_tensors[self.fold_columns[name]].append(id_tensor)
-
-    for fold_name, id_tensor in id_tensors.items():
-      id_tensors_concat = tf.concat(id_tensor, axis=1)
-      embedding_out_concat = self.embedding_layers[fold_name](id_tensors_concat)
-      embedding_out = tf.split(embedding_out_concat, num_or_size_splits=self.split_dims_final[fold_name], axis=1)
-
-      for i, embedding in enumerate(embedding_out):  # LENGTH(embedding_out) == split_dims_final
-        if self.is_sequence_feature[fold_name][i] and True:
-          embedding = tf.math.reduce_mean(embedding, axis=1, keepdims=True)  # (feature_combin_num, (batch, x, 16*6+1))
-        result[fold_name].append(embedding)
-
-    return result
 
 
 class QREmbedding(tf.keras.layers.Layer):
