+++ deepray/version.py
@@ -15,20 +15,20 @@
 """Define Deepray version information."""
 
 # Required TensorFlow version [min, max)
-INCLUSIVE_MIN_TF_VERSION = "2.10.0"
+INCLUSIVE_MIN_TF_VERSION = "2.9.3"
 EXCLUSIVE_MAX_TF_VERSION = "2.13.0"
 
 # We follow Semantic Versioning (https://semver.org/)
 _MAJOR_VERSION = "0"
 _MINOR_VERSION = "21"
-_PATCH_VERSION = "0"
+_PATCH_VERSION = "1"
 
 # When building releases, we can update this value on the release branch to
 # reflect the current release candidate ('rc0', 'rc1') or, finally, the official
 # stable release (indicated by `_VERSION_SUFFIX = ''`). Outside the context of a
 # release branch, the current version is by default assumed to be a
 # 'development' version, labeled 'dev'.
-_VERSION_SUFFIX = "dev"
+_VERSION_SUFFIX = ""
 
 # Example, '0.1.0-dev'
 __version__ = ".".join([_MAJOR_VERSION, _MINOR_VERSION, _PATCH_VERSION])
